Q9 is completed

+++ DSA/Operators/ArithmeticO/Q9.java
@@ -2,7 +2,7 @@ package DSA.Operators.ArithmeticO;
 
 public class Q9 {
     public static void main(String[] args) {
-        int x = 5/2*2;
+        double x = 5/2*2;
         System.out.println(x);
     }
 }
